% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-annotations.R
\name{annotation_textbox}
\alias{annotation_textbox}
\title{HTML textbox annotation}
\usage{
annotation_textbox(
  html,
  position = "top-left",
  width = NULL,
  height = NULL,
  x_offset = 0,
  y_offset = 0,
  gp = grid::gpar(fontsize = 10, lineheight = 1.4, fontfamily = "Inter-Medium"),
  box_gp = grid::gpar(col = "#0E151B", fill = "#e5e7eb", lwd = grid::unit(1.5, "pt")),
  padding = grid::unit(rep(10, 4), "pt"),
  margin,
  r = grid::unit(5, "pt"),
  orientation = "upright",
  ...
)
}
\arguments{
\item{html}{Content of the textbox in html syn \%>\% tax. See supported tags in \url{https://wilkelab.org/ggtext/index.html}.}

\item{position}{One of "top-left", "top-right", "bottom-left", "bottom-right", "center". Calculated with \code{compute_position}.}

\item{width}{A numeric value for the npc or a \code{grid::unit} object. Defaults to fit the contents passed to \code{html}.}

\item{height}{A numeric value for the npc or a \code{grid::unit} object. Defaults to fit the contents passed to \code{html}.}

\item{x_offset}{x-position offset relative to the panel space.}

\item{y_offset}{y-position offset relative to the panel space.}

\item{gp}{Graphical parameters for the text as \code{grid::gpar}.}

\item{box_gp}{Graphical parameters for the bounding box as \code{grid::gpar}.}

\item{padding}{Spacing for the inner-space between the box boundaries and the box contents, supplied as \code{grid::unit}.}

\item{margin}{Spacing for the outer-space between the box boundaries and its position, supplied as \code{grid::unit}.}

\item{r}{Radius of the box corners, supplied as \code{grid::unit}.}

\item{orientation}{Orientation of the box. One of "upright", "left-rotated", "right-rotated", "inverted".}

\item{...}{Passed to \code{gridtext::textbox_grob}}
}
\value{
A \code{gridtext::textbox_grob}.
}
\description{
A wrapper around \code{gridtext::textbox_grob} with convenient positioning and themed defaults.
}
\details{
To be used for non-data-based annotations with \code{ggplot2::annotation_custom}.
}
\examples{
\dontrun{
annotation <- annotation_textbox(
  html = "<span style=\"font-size:18px; font-family:'Inter-Black';\">
           This is a cat:</span>
          <br><br>
          <span style=\"font-size:72px; color:'#7950F2';
           font-family:'FontAwesome5-Solid';\">cat</span>
          <br><br>
          Just look at this cat!",
  position = "top-right",
  halign = 0.5
)

ggplot2::qplot(mpg, hp, data = mtcars) +
  ggplot2::annotation_custom(annotation) +
  ggplot2::theme_classic()
}
}
