% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-annotations.R
\name{compute_position}
\alias{compute_position}
\title{Return npc-based positional values for common alignments}
\usage{
compute_position(position)
}
\arguments{
\item{position}{One of "top-left", "top-right", "bottom-left", "bottom-right", "center".}
}
\value{
A list of \code{hjust}, \code{vjust}, \code{x}, and \code{y}
}
\description{
Return npc-based positional values for common alignments
}
