% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-load.R
\name{.pgl_fonts}
\alias{.pgl_fonts}
\title{Fonts included in the package}
\usage{
.pgl_fonts()
}
\value{
A named list of font families and font styles
}
\description{
Fonts included in the package
}
\examples{
\dontrun{
.pgl_fonts()
}
}
