% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-factors.R
\name{fct_match}
\alias{fct_match}
\title{Derive factor levels from another factor with unique pair-wise matches}
\usage{
fct_match(new, old)
}
\arguments{
\item{new}{A vector}

\item{old}{A factor}
}
\value{
A factor
}
\description{
Derive factor levels from another factor with unique pair-wise matches
}
\examples{
old <- factor(LETTERS[1:5], LETTERS[c(3, 2, 1, 5, 4)])
old
new <- letters[1:5]
new
factor(new)
fct_match(new, old)
}
