% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{font_exists}
\alias{font_exists}
\title{Check whether a font is registered}
\usage{
font_exists(family, style = "Regular")
}
\arguments{
\item{family}{Font family.}

\item{style}{Font style. Defaults to \code{"Regular"}.}
}
\value{
Boolean
}
\description{
Wrapper around \code{font_info} that tests whether the output is \code{NULL}.
}
\examples{
\dontrun{
font_exists(family = "Arial")
}
}
