% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-helpers.R
\name{font_hoist}
\alias{font_hoist}
\title{Register members of a font family as their own families}
\usage{
font_hoist(family, verbose = TRUE)
}
\arguments{
\item{family}{The name of the font family, as registered in \code{systemfonts::system_fonts()}}

\item{verbose}{Whether to print a message of the newly registered font families}
}
\description{
Register members of a font family as their own families
}
\examples{
\dontrun{
font_hoist(family = "Arial")
}

}
