% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{font_info}
\alias{font_info}
\title{Get information about a font}
\usage{
font_info(family, style = "Regular")
}
\arguments{
\item{family}{Font family.}

\item{style}{Font style. Defaults to \code{"Regular"}.}
}
\value{
A dataframe containing information about a font.
The \code{"source"} attribute indicates where it was found.
If the font doesn't exist, invisibly returns \code{NULL}.
}
\description{
Get information about a font
}
