% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{font_path}
\alias{font_path}
\title{Find path to a system font}
\usage{
font_path(family, style = "Regular")
}
\arguments{
\item{family}{Font family.}

\item{style}{Font style. Defaults to \code{"Regular"}.
If the font doesn't exist, invisibly returns \code{NULL}.}
}
\value{
String of the path to the font file.
}
\description{
Wrapper around \code{font_info} that pulls the path of the font if it exists.
}
