% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-helpers.R
\name{font_pluck}
\alias{font_pluck}
\title{Register a font variant with OpenType feature specifications}
\usage{
font_pluck(family, features, .name)
}
\arguments{
\item{family}{A font family. The Regular member of the family is used as the base.}

\item{features}{Character vector of 4-letter OpenType feature tags.
See \href{https://docs.microsoft.com/en-us/typography/opentype/spec/featurelist}{complete list of OpenType features}.
Unavailable font features are ignored.}

\item{.name}{(Optional) Name of the new font family. Appends registered feature tags to the new family name by default.}
}
\description{
Register a font variant with OpenType feature specifications
}
\examples{
\dontrun{
font_pluck(family = "Alegreya", c("lnum", "ordn"))
}

}
