% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-save-render.R
\name{ggsave2}
\alias{ggsave2}
\title{Save and open plot}
\usage{
ggsave2(...)
}
\arguments{
\item{...}{Passed to \code{ggplot2::ggsave()}}
}
\description{
Wrapper around \code{ggplot2::ggsave()} that opens rendered output using the system's default app for the graphic type.
}
