% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-save-render.R
\name{ggsave_auto}
\alias{ggsave_auto}
\title{Save and open plot}
\usage{
ggsave_auto(..., load_as_magick = FALSE)
}
\arguments{
\item{...}{Passed to \code{ggplot2::ggsave()}, overriding \code{path} and \code{filename}.}

\item{load_as_magick}{Whether to show information about the saved plot and invisibly return an ImageMagick object of the plot for post-processing. Defaults to \code{FALSE}.}
}
\value{
An magick object or path to the saved plot.
}
\description{
Wrapper around \code{ggplot2::ggsave()} that saves the rendered output to a special folder
and opens it using the system's default app for the graphic type.
}
\details{
Creates a folder \verb{/.ggsave_auto} in the working directory where figures are saved
with zero-padding to 3 digits (e.g., \verb{ggsave_auto-001.png}).
}
