% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-load.R
\name{load_pkg_fonts}
\alias{load_pkg_fonts}
\title{Load package fonts}
\usage{
load_pkg_fonts(verbose = TRUE)
}
\arguments{
\item{verbose}{Whether the newly registered fonts should be printed to the console. Defaults to \code{FALSE}.}
}
\description{
Registers fonts contained in {penngradlings} with \code{systemfonts::register_font()}.
Call \code{systemfonts::registry_fonts()} after loading the fonts for more details.
}
\section{Fonts loaded}{

\describe{
\item{\verb{Charis SIL}}{ From SIL International \url{https://software.sil.org/charis/}}
\item{\code{Inter}}{ From Google Fonts \url{https://fonts.google.com/specimen/Inter}}
\item{\code{Lato}}{ From Google Fonts \url{https://fonts.google.com/specimen/Roboto}}
}

All fonts are licensed under the SIL Open Font License (OFL) \url{https://scripts.sil.org/OFL}
}

