% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{mean_se_corrected}
\alias{mean_se_corrected}
\title{Corrected calculation of mean and standard error}
\usage{
mean_se_corrected(x, mult = 1)
}
\arguments{
\item{x}{Numeric vector.}

\item{mult}{Number of multiples of standard error.}
}
\value{
A dataframe with three columns:
\describe{
\item{\code{y}}{ The mean.}
\item{\code{ymin}}{ The mean minus the multiples of the standard error.}
\item{\code{ymax}}{ The mean plus the multiples of the standard error.}
}
}
\description{
Adapted from \code{ggplot2::mean_se}. Corrects the denominator of the variance with \eqn{N - 1}.
}
\examples{
x <- rnorm(100)
mean_se_corrected(x)
}
