% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metaprogramming.R
\name{name_self}
\alias{name_self}
\title{Construct a list that names itself}
\usage{
name_self(list_expr, .fn = function(x, y) {
    x
})
}
\arguments{
\item{list_expr}{A expression using a call to \code{list}, \code{c}, or the like, where all arguments
are name-value pairs and the expression evaluates to a list.}

\item{.fn}{A function passed to \code{purrr::imap_chr}, called on the character vector returned
after coercing the element symbols in \code{list_expr} to name. Defaults to the identity function.}
}
\value{
A named list
}
\description{
Construct a list that names itself
}
\examples{
name_self(c(min, mean, max))
name_self(c(min, mean, max), .fn = ~ toupper(.x))
name_self(c(min, mean, max), .fn = function(x, y) {
  paste0(x, y)
})

# Helps when naming produced by from `across()`
suppressPackageStartupMessages(library(dplyr))

## With `name_self()`, columns are named after the functions
mtcars \%>\%
  group_by(cyl) \%>\%
  summarize(across(where(~ max(.x) > 100), name_self(list(min, mean, max))))

## Without `name_self()`, uses position indices
mtcars \%>\%
  group_by(cyl) \%>\%
  summarize(across(where(~ max(.x) > 100), list(min, mean, max)))
}
