% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penngradlings-package.R
\docType{package}
\name{penngradlings-package}
\alias{penngradlings}
\alias{penngradlings-package}
\title{penngradlings: Themes and Utilities for Penn GradLingS}
\description{
Assortment of quality-of-life functionalities for common workflows in R, centered around data visualization and science communication. Made by and for the University of Pennsylvania Graduate Linguistics Society (GradLingS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yjunechoe.github.io/penngradlings}
  \item \url{https://github.com/yjunechoe/penngradlings}
  \item Report bugs at \url{https://github.com/yjunechoe/penngradlings/issues}
}

}
\author{
\strong{Maintainer}: June Choe \email{yjchoe@sas.upenn.edu} (\href{https://orcid.org/0000-0002-0701-921X}{ORCID})

}
\keyword{internal}
