% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrangling-misc.R
\name{qcut}
\alias{qcut}
\title{Discretize a continuous variable into quantiles}
\usage{
qcut(x, q = c(0.25, 0.5, 0.75), ...)
}
\arguments{
\item{x}{Numeric}

\item{q}{Probabilities. Should be 1 less than the number of discrete bins to be returned.
\code{x} is cut from left-to-right, filling up the smallest bin first. For example, \code{c(.25, .50, .75)}
returns a factor of 4 levels - 0\%-25\%, 25\%-50\%, 50\%-75\%, 75\%-100\%. Bins are left-inclusive and right-exclusive.}

\item{...}{Passed to \code{stats::quantile}}
}
\description{
Discretize a continuous variable into quantiles
}
\examples{
qcut(1:100, c(.2, .4, .6, .8))
}
