% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-save-render.R
\name{raggsave_auto}
\alias{raggsave_auto}
\title{Save and open any plot}
\usage{
raggsave_auto(plot_expr = NULL, ..., load_as_magick = FALSE)
}
\arguments{
\item{plot_expr}{An expression that plots a figure to a device as a side effect, to be captured by \code{ragg::agg_png()}.}

\item{...}{Passed to \code{ragg::agg_png()}, overriding \code{filename}. Uses the same width and height as the RStudio viewer pane size by default, like \code{ggplot2::ggsave()}.}

\item{load_as_magick}{Whether to show information about the saved plot and invisibly return an ImageMagick object of the plot for post-processing. Defaults to \code{FALSE}.}
}
\value{
A magick object or path to the saved plot.
}
\description{
Wrapper around \code{ragg::agg_png()} that saves the output from a plotting expression to a special folder
and opens it using the system's default app for the graphic type.
}
\note{
Creates a folder \verb{/.raggsave_auto} in the working directory where figures are saved
with zero-padding to 3 digits (e.g., \verb{raggsave_auto-001.png}). Appends \code{.raggsave_auto} to
a \code{.gitignore} file if it exists.

Global params can be set through \code{options()}, with the argument name followed by \code{pgl.raggsave_auto.}
such as \code{options(pgl.raggsave_auto.width = 7)}.
}
