% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_pcibex}
\alias{read_pcibex}
\title{Read the results from a PCIbex experiment}
\usage{
read_pcibex(file, exclude_controller_name = TRUE)
}
\arguments{
\item{file}{Path to the results file}

\item{exclude_controller_name}{Whether to skip parsing a column for \code{"Controller name"}.
In the new PCIbex (last checked: September 2021), \code{"Controller name"} appears in the column
specification as the third column but does not have associated values in the results. This is
\code{TRUE} by default but when parsing results files from the old PCIbex, it should be set to \code{FALSE}.}
}
\value{
A dataframe
}
\description{
Converts a comma-separated PCIbex results file into rectangular format.
}
\examples{
\dontrun{
# It takes a few seconds because it's reading from remote
# but the parsing itself is very fast.
# Also note that this results file is from the old PCIbex
# so `exclude_controller_name` is set to `FALSE`
dplyr::glimpse(
  read_pcibex(
    "https://raw.githubusercontent.com/yjunechoe
     /Semantic-Persistence/master/data/result.txt",
     exclude_controller_name = FALSE
  )
)
}
}
