% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrangling-misc.R
\name{rshuffle}
\alias{rshuffle}
\title{(Randomized) even shuffling of categories}
\usage{
rshuffle(x, mode = c("squish", "symmetric", "repel"), shuffle_within = FALSE)
}
\arguments{
\item{x}{Vector}

\item{mode}{Algorithm for even shuffling:
\itemize{
\item "squish" (default): Within each category, maximize the distance between observations, away from the ends of the vector.
\item "symmetric": Maximize distance around the category with the smallest number of observations.
\item "repel": Within each category, maximize the distance between observations, period.
}}

\item{shuffle_within}{Whether observations within each category should be shuffled.
Defaults to \code{FALSE}, which keeps the ordering of observations within each category.}
}
\value{
A vector of shuffled indices of \code{x}
}
\description{
(Randomized) even shuffling of categories
}
\examples{

ex <- c(rep("a", 10), rep("bb", 5), rep("ccc", 3), rep("a", 2), rep("bb", 3), rep("ccc", 3))
ex

# Demos of algorithms
barplot(nchar(ex[rshuffle(ex)])) # mode = "squish"
barplot(nchar(ex[rshuffle(ex, mode = "symmetric")]))
barplot(nchar(ex[rshuffle(ex, mode = "repel")]))

# Returns indices for ease of reordering rows
data.frame(id = LETTERS, category = ex)[rshuffle(ex), , drop = FALSE]

}
