% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-colors.R
\name{scale_color_pgl_discrete}
\alias{scale_color_pgl_discrete}
\alias{scale_colour_pgl_discrete}
\alias{scale_fill_pgl_discrete}
\title{Discrete color scales}
\usage{
scale_color_pgl_discrete(what = NULL, who = NULL, ...)

scale_colour_pgl_discrete(what = NULL, who = NULL, ...)

scale_fill_pgl_discrete(what = NULL, who = NULL, ...)
}
\arguments{
\item{what, who}{passed to \code{pgl_pals()}}

\item{...}{passed to \code{ggplot2::discrete_scale()}}
}
\description{
Discrete color scales
}
