% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinytable.R
\name{table_lm}
\alias{table_lm}
\title{Quick table of regression model output}
\usage{
table_lm(x)
}
\arguments{
\item{x}{A linear model with a \code{broom::tidy()} method defined}
}
\value{
A \code{tinytable }table
}
\description{
Quick table of regression model output
}
\examples{
x <- lm(mpg ~ hp, mtcars)
table_lm(x) # \%>\% tinytable::save_tt("ex.docx") \%>\% browseURL()
}
