% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-themes.R
\name{theme_pgl_minimal}
\alias{theme_pgl_minimal}
\title{Basic {penngradlings} theme template}
\usage{
theme_pgl_minimal(
  base_size = 12,
  axis_lines = "",
  grid_lines = "xy",
  grid_lines_minor = FALSE,
  outline_color = "#1F1C1C"
)
}
\arguments{
\item{base_size}{Base size for text elements}

\item{axis_lines}{Which axis lines should be drawn, passed as a string.}

\item{grid_lines}{Which grid lines should be drawn, passed as a string.}

\item{grid_lines_minor}{Whether minor grid lines should be draw. Defaults to \code{FALSE}}

\item{outline_color}{Primary color for text, line, and rectangle outline.}
}
\value{
A theme object created by \code{theme()}
}
\description{
Basic {penngradlings} theme template
}
\examples{
\dontrun{
library(ggplot2)

ggplot(iris, aes(Petal.Length, Petal.Width, color = Species)) +
  geom_point(size = 2) +
  labs(
    title = "Petal profile of plant species in the iris dataset",
    subtitle = "Virginica has the largest petal of the three species"
  ) +
  facet_wrap(~Species) +
  theme_pgl_minimal(base_size = 11)

ggplot(diamonds, aes(cut, price, fill = cut)) +
  stat_summary(geom = "bar", fun = "mean", width = 0.8) +
  scale_y_continuous(
    expand = expansion(c(0, 0.1)),
    labels = ~ paste0("$", .x)
  ) +
  scale_fill_manual(values = pgl_pals("christine", "blueberry_boba")) +
  labs(
    title = "Price of diamonds by cut quality",
    subtitle = "This is a bad explanatory plot because carat (size) is a confounding variable",
    x = NULL, y = NULL,
    caption = "diamonds dataset from {ggplot2}"
  ) +
  theme_pgl_minimal(axis_lines = "x", grid_lines = "y")
}
}
