% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-themes.R
\name{theme_plg_basic}
\alias{theme_plg_basic}
\title{Basic {penngradlings} theme template}
\usage{
theme_plg_basic(
  base_size = 12,
  axis_lines = "",
  grid_lines = "xy",
  grid_lines_minor = FALSE
)
}
\arguments{
\item{base_size}{Base size for text elements}

\item{axis_lines}{Which axis lines should be drawn, passed as a string.}

\item{grid_lines}{Which grid lines should be drawn, passed as a string.}

\item{grid_lines_minor}{Whether minor grid lines should be draw. Defaults to \code{FALSE}}
}
\value{
A theme object created by \code{ggplot2::theme()}
}
\description{
Basic {penngradlings} theme template
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(mpg, hp)) +
  geom_point(aes(color = factor(cyl))) +
  labs(
    title = "This is an example plot",
    subtitle = "We are using this to demonstrate theme_plg_basic",
    x = "Miles per gallon (mpg)",
    y = "Horsepower",
    color = "Cylinder"
  ) +
  theme_plg_basic()
}
}
