% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-defaults.R
\name{update_ggplot_defaults}
\alias{update_ggplot_defaults}
\title{Modify defaults ggplot2 layer aesthetics for {penngradlings}}
\usage{
update_ggplot_defaults()
}
\description{
Modify defaults ggplot2 layer aesthetics for {penngradlings}
}
