% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-annotations.R
\name{zerowidth_char}
\alias{zerowidth_char}
\title{Create an HTML span of the unicode zero-width character}
\usage{
zerowidth_char(size = 12, units = "px")
}
\arguments{
\item{size}{Size of the zero-width character}

\item{units}{Defaults to \code{"px"}}
}
\value{
A string
}
\description{
Helper function for adjusting paragraph margins for use with {ggtext}/{gridtext},
which does not natively support margin/padding styles.
}
\details{
The zero-width unicode character only ever takes up vertical space, so when placed in-line
its size determines the amount of "top-margins" for the paragraph. This allows finer control
over spacing.
}
\examples{
\dontrun{
library(ggtext)

ggplot() +
  geom_richtext(aes(0, 0, label = "This is a line<br><br>This is a paragraph<br>of text."))

ggplot() +
  geom_richtext(aes(0, 0, label = paste0("This is a line<br>",
                                         zerowidth_char(30),
                                         "This is a paragraph<br>of text.")))
}
}
