% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{font_hoist}
\alias{font_hoist}
\title{Register members of a font family as their own families}
\usage{
font_hoist(family, silent = FALSE)
}
\arguments{
\item{family}{The name of the font family, as registered in \code{systemfonts::system_fonts()}}

\item{silent}{Whether to print a message of the newly registered font families}
}
\description{
Register members of a font family as their own families
}
\examples{
font_hoist(family = "Arial")

}
