% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-annotations.R
\name{geom_text_outline}
\alias{geom_text_outline}
\title{Outlined text layer}
\usage{
geom_text_outline(
  ...,
  geom = "geom_text",
  inner_params = list(),
  outer_params = list(),
  use_outer = TRUE
)
}
\arguments{
\item{...}{Passed to the geom layer as specified in the \code{geom} argument.}

\item{geom}{Which geom layer to apply outlines over. Defaults to}

\item{inner_params}{Passed to the first \code{ggfx::with_outer_glow} filter (inner outline).}

\item{outer_params}{Passed to the second \code{ggfx::with_outer_glow} filter (outer outline).}

\item{use_outer}{Whether the outer outline should be turned on. Defaults to \code{TRUE}.}
}
\value{
A {ggplot2} layer if \code{use_outer} is \code{TRUE}, otherwise a list of two layers.
}
\description{
Mimicking SVG-style "stroke" styles, this function uses 1-2 \code{ggfx::with_outer_glow} filters
to create outlines around text. There are two layers for the following uses: the first layer is designed to
create a visible outline that places emphasis on the text. The second layer is designed to add
a mask around the outlined text that makes it stand out from the background.
}
\details{
This is particularly useful for labels drawn over panels with grid lines and a non-white background color.
Outlines will create a stronger perceptual contrast for the labels.
}
\examples{
\dontrun{
library(ggplot2)
p <- ggplot(mtcars, aes(mpg, disp, label = rownames(mtcars)))

# By default creates a white inner-outline and a gray-shade outer-outline
p + geom_text_outline()

# This is more obvious against a white background
p + geom_text_outline() +
  theme_void()

# You also have the option to turn the outer outline off
p + geom_text_outline(use_outer = FALSE) +
  theme_void()

# You can pass arguments to the layer specified in the `geom` argument in the `...`.
# The default geom is `geom_text`, so you set values for `check_overlap` and `angle`
p + geom_text_outline(check_overlap = TRUE, angle = 30)

# You can invert the text outlines by setting the color of the text to "white"
# and the color of the inner outline to "black".
p + geom_text_outline(color = "white", inner_params = list(colour = "black"), use_outer = FALSE) +
  theme_void()

# You can pass other geoms to the `geom` argument
library(ggrepel)
p + geom_text_outline(geom = geom_text_repel)
}
}
