% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-load.R
\name{load_pkg_fonts}
\alias{load_pkg_fonts}
\title{Load package fonts}
\usage{
load_pkg_fonts(verbose = TRUE)
}
\arguments{
\item{verbose}{Whether the newly registered fonts should be printed to the console. Defaults to \code{TRUE}.}
}
\value{
An invisible call to \code{systemfonts::registry_fonts()} for further inspection.
}
\description{
Registers fonts contained in {penngradlings} with \code{systemfonts::register_font()}. This function is called once silently when the package is loaded.
}
\details{
All styles of a font family are registered as the Regular member of their own families in the form of "{family}-{style(s)}".

Call \code{systemfonts::registry_fonts()} after loading the fonts for more details (also returned invisibly).
}
\section{Fonts loaded}{

\describe{
\item{\verb{Charis SIL}}{ by SIL international, retrieved from \url{https://software.sil.org/charis/}. }
\item{\code{Inter}}{ by Rasmus Andersson, retrieved from \url{https://fonts.google.com/specimen/Inter}. }
\item{\code{Piazzolla}}{ by Juan Pablo del Peral, retrieved from \url{https://fonts.google.com/specimen/Piazzolla}. }
\item{\verb{Bootstrap Icons}}{ by The Bootstrap Authors, retrieved from \url{https://icons.getbootstrap.com/}. }
\item{\verb{Material Icons}}{ by Google Design, retrieved from \url{https://fonts.google.com/icons/}. }
\item{\verb{Font Awesome 5 Free}}{ by Dave Gandy \url{https://fontawesome.com/}. }
The free version contains a subset of the Regular, Solid, and Brand styles:
\url{https://fontawesome.com/v5.15/icons?d=gallery&p=2&s=brands,regular,solid&m=free}.
}

Copy of the license for each font is included in the package source.
The use of these fonts in the package does not indicate endorsement or liability of the copyright/trademark holders.
We thank the developers of the fonts for their open-source contributions and permissive licenses.
}

\examples{
\dontrun{
load_pkg_fonts()
}

}
