% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-system.R
\name{nav}
\alias{nav}
\title{Navigate to a file}
\usage{
nav(x)
}
\arguments{
\item{x}{A path.}
}
\value{
Invisibly returns \code{x}.
}
\description{
When used interactively, opens the file/folder.
}
\examples{
\dontrun{
if (interactive()) {
  nav(getwd())
 }
}
}
