% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-colors.R
\name{pgl_pals}
\alias{pgl_pals}
\title{Select a PGL-themed color palette}
\usage{
pgl_pals(who, what, n)
}
\arguments{
\item{who}{The person associated with the palette.}

\item{what}{The object the palette was extracted from}

\item{n}{Number of colors to return}
}
\value{
A character vector of colors
}
\description{
Select a PGL-themed color palette
}
\examples{
pgl_pals("christine", "blueberry_boba", 4)
}
