% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_pcibex}
\alias{read_pcibex}
\title{Read the results from a PCIbex experiment}
\usage{
read_pcibex(file, encoding = "UTF-8", exclude_controller_name = FALSE)
}
\arguments{
\item{file}{Path to the results file}

\item{encoding}{The name of the encoding to be assumed.}

\item{exclude_controller_name}{Whether to skip parsing a column for \code{"Controller name"}.
In a dev version of PCIbex (last checked: September 2021), \code{"Controller name"} appears in the column
specification as the third column but does not have associated values in the results. This has
since been fixed, so the value is set to \code{FALSE} by default.}
}
\value{
A dataframe
}
\description{
Converts a comma-separated PCIbex results file into rectangular format.
}
\examples{
\dontrun{
# It takes a few seconds because it's reading from remote
# but the parsing itself is very fast.
dplyr::glimpse(
  read_pcibex(
    "https://raw.githubusercontent.com/yjunechoe
     /Semantic-Persistence/master/data/result.txt",
    exclude_controller_name = FALSE
  )
)
}
}
