% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_pcibex}
\alias{read_pcibex}
\title{Read the results from a PCIbex experiment}
\usage{
read_pcibex(file)
}
\arguments{
\item{file}{Path to the results file}
}
\value{
A dataframe
}
\description{
Converts the comma-separated PCIbex results file into rectangular format.
}
