% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_pcibex}
\alias{read_pcibex}
\title{Read the results from a PCIbex experiment}
\usage{
read_pcibex(file)
}
\arguments{
\item{file}{Path to the results file}
}
\value{
A dataframe
}
\description{
Converts a comma-separated PCIbex results file into rectangular format.
}
\examples{
\dontrun{
# Takes a few seconds because it's reading from remote
# but the parsing itself is very fast
read_pcibex("https://raw.githubusercontent.com/yjunechoe
             /Semantic-Persistence/master/data/result.txt")
}
}
