% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_pcibex}
\alias{read_pcibex}
\title{Read the results from a PCIbex experiment}
\usage{
read_pcibex(
  file,
  encoding = "UTF-8",
  exclude_controller_name = FALSE,
  type_opts = list()
)
}
\arguments{
\item{file}{Path to the results file}

\item{encoding}{The name of the encoding to be assumed.}

\item{exclude_controller_name}{For backwards compatibility, whether to skip the column for \code{"Controller name"}.
In a previous version of PCIbex (last checked: September 2021), \code{"Controller name"} appeared in the column
specification as the third column but does not have associated values in the results. This has
since been fixed, so the value is set to \code{FALSE} by default.}

\item{type_opts}{A list of arguments passed to \code{type.convert()}, or \code{readr::type_convert()} if tidyverse is installed.}
}
\value{
A dataframe
}
\description{
Converts a comma-separated PCIbex results file into rectangular format.
}
