% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinytable.R
\name{table_lm}
\alias{table_lm}
\title{Quick table of regression model output}
\usage{
table_lm(x, beta_digits = 2, write = FALSE)
}
\arguments{
\item{x}{A model with a \code{broom::tidy()} method defined, or pre-tidied data.}

\item{beta_digits}{Digits to approximate on the estimate. This also controls
digits printed for the S.E. and t-statistic columns. Defaults to \code{2}.}

\item{write}{Whether to write to a temporary file and open it in Viewer.
Defaults to \code{FALSE}. Can specify a format for \code{tinytable::save_tt()}.}
}
\value{
A \code{tinytable }table
}
\description{
Quick table of regression model output
}
\examples{
x <- lm(mpg ~ hp, mtcars)
table_lm(x) # `write = ".docx"` to also write out to a Word file
}
